+++ src/Collector/TemplateAssignment/Data/TypeInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PHPStanConfig\Collector\TemplateAssignment\Data;
+
+interface TypeInterface
+{
+    public function describe(): string;
+}