+++ src/Collector/TemplateAssignment/Data/Types/IntegerType.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PHPStanConfig\Collector\TemplateAssignment\Data\Types;
+
+use PHPStanConfig\Collector\TemplateAssignment\Data\TypeInterface;
+
+final class IntegerType implements TypeInterface
+{
+    /** @implements SetStateTrait<IntegerType> */
+    use SetStateTrait;
+
+    public function __construct()
+    {
+        $this->__myClass = self::class;
+    }
+
+    public function describe(): string
+    {
+        return 'int';
+    }
+}