+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- added rules and collectors for plugin bridge data, template assignment code and application/plugin configs
 
 ## [0.19.7] - 2025-01-12
 ### Fixed